Add support for UINavigationItem
This change adds support for setting a `UINavigationItem` on a `HUBViewModel`,
which will then be used when its `HUBViewController` is added to a navigation
controller.

This is done by replacing `navigationBarTitle` with `navigationItem` on
`HUBViewModel` (since you can now just use the navigation item’s title).
No backwards compatibility is added here, since the use cases for using a
`HUBViewModel` directly are very small.

However, on the builder (`HUBViewModelBuilder`) backwards compatibility is
achieved by letting `navigationBarTitle` be an alias for `navigationItem.title`.

Some equality checking code is extracted from `HUBAutoEquatable` into `HUBUtilities`
to avoid code duplication.