Set the viewModel after performing batch updates
UICollectionView’s “performBatchUpdates:” method checks the number of
items in the collection view before performing the update block and
after that. It expects the following: “The number of items contained in
an existing section after the update must be equal to the number of
items contained in that section before the update, plus or minus the
number of items inserted or deleted from that section and plus or minus
the number of items moved into or out of that section.”. If we update
the view model before performing the batch updates
“collectionView:numberOfItemsInSection:” will return the updated count
even before the actual update and it will fail after that with uncaught
exception 'NSInternalInconsistencyException', reason: 'Invalid update’.
For example we have 0 items initially and we add 2, before the block we
will have 2 items and we will add 2 from the “lastViewModelDiff” so at
the end the collection view will expect to have 4 items but it will
have only 2.