Add API to create view controllers without registering features
This change adds a new way to create view controllers, without having to
pre-register a feature with the framework. The use case for this is to
more easily be able to create view controller, which decreases the learning
curve to get into using the framework.

It also will make it easier to use the Hub Framework for playgrounds.

To make this happen, another convenience API is introduced to be able to
create content operation factories using a concrete class that takes a block.
Again, this reduces the barrier of entry for creating content operations, and
the need for boilerplate.