Support nested touchable views inside a UI[Table/Collection]ViewCell
Because of the ways UIKit handles touches in `UITableViewCell` and
`UICollectionViewCell`, and we want to support using “orphaned” cells
as component views - we need to employ some custom touch handling.

It turns out though, that the custom touch handling wasn’t taking nested
touchable views (such as a `UIButton` inside a cell) into account. This
has now been fixed, by making sure we only do custom handling when the
content view of a cell was touched.