Make highlight & selection handling more dynamic
This change makes the way components are highlighted & selected a lot more
dynamic. Before, only components implemented using either `UITableViewCell`
or `UICollectionViewCell` would be able to be have highlight & selection
appearances (because we simply forwarded those events from the wrapper cell).

With this change, a `HUBComponentWithSelectionState` protocol is introduced,
making any component able to respond to selection state changes. Selection &
highlighting is now also done through a custom gesture recognizer, instead of
attempting to hack UIKit’s touch handling code (turned out to be a mistake).

The end goal of this change is to:

1) Enable any component to be highlighted & selected.
2) Simplify the implementation of selection handling.
3) Enable components to have custom gesture recognizers.