Support highlight & selection out of the box for table & collection view cells
This re-adds the previous behavior of automatically highlighting and selecting
components that are implemented using either a table- or collection view cell.

Even though the new `HUBComponentWithSelectionState` protocol enables any view
to be highlighted and selected - this new mechanism makes it a lot more convenient
when implementing components using cells, and avoids code duplication in having
to handle those selection states manually.