View controller: Un-rename visible component view’s API & improve docs
This change changes the name of the API used to retrieve the currently
visible body component views back to the name that was used before
https://github.com/spotify/HubFramework/pull/84. We don’t need to change
the name of the API because the key type was changed.

This change also adds a bit more thorough documentation for the API,
explaining what the API user can expect from the index paths used as keys.