Demo app: Add carousel component
This change adds a carousel component to the demo app, partly for demonstration
purposes - but also to enable us to cheaply perform UI tests and manual verification
for a component with children (the demo app previously didn’t have any components
with children).

The “Pretty pictures” feature of the demo app is also updated to display a carousel
on the top, as well as adding tappable links to the pictures that it displays.