View controller: Send layout traits for component appear/disappear
This change makes `HUBViewController` send a component’s layout traits
to its delegate as part of the methods called whenever a component is
either about to appear, or just disappeared from the screen.

The reason for this change is that we want to enable API users to make
decisions based on the layout traits, instead of resorting to comparing
view frames (which are bound to change during the layout cycle).