View model loader: Support appending paginated content
This change introduces an API for appending paginated content to a view model.
By making a content operation conform to `HUBContentOperationWithPaginatedContent`,
new content can now be added to an existing view model builder.

To support the declarative nature of the Hub Framework, content operations are
always guaranteed to execute from a “clean state”. `HUBViewModelLoader` keeps snapshots
of previous view model builders & errors, to be able to make this guarantee.

However, with pagination, we don’t want to execute from a clean state. Instead, we
want to append content to the **current** state, which imposes some interesting
challenges on the content loading code.

To accommodate these 2 different uses cases, a new `HUBContentOperationExecutionInfo`
is introduced. Instead of enqueueing operations directly, instances of this class
is now used to keep track of a view model loader’s operation queue. This enables us
to reference the same content operation using different modes, and to use different
snapshots depending on what mode an operation is being executed in.

Went all in with the tests this time, to make sure that A) this doesn’t break any
existing content loading code, and B) that all the edge cases around managing both
replacing and appended state changes at the same time.