Demo app: Add a “sticky header” component + feature and use it for UI tests
This change adds a header component that applies a parallax-like effect to
its background image. It’s showcased in a new “Sticky header” feature.

This is done partly for demonstration purposes, but mostly to be able to
perform UI tests against a header component, to assert that the framework
behaves the correct way when one is used.

To avoid further code duplication in UI tests, a `UITestCase` base class is
created that performs the “boilerplate” operations involved in setting up
a UI test.