Use HUBActionHandlerWrapper as HUBActionHandler protocol
This change is part of the work for https://github.com/spotify/HubFramework/issues/158,
which has the end goal of making `HUBViewController` publicly initializable. To do this,
we need to remove all dependencies on concrete internal classes.

This is the first step, which makes `HUBViewController` take any `HUBActionHandler` as
part of its initializer, rather than relying on the concrete `HUBActionHandlerWrapper`
implementation. To make this happen, logic that was previously delegated to the view
controller around chaining of async actions needed to be moved into the action handler
wrapper itself.

The test change that was made in `HUBViewControllerTests` needed to make made to avoid
implicit autorelease pools from taking over the lifecycle of objects.