Expose HUBViewController as a concrete class
This change makes `HUBViewController` a concrete class (instead of an API
protocol/implementation class pair) and exposes it as part of the public
API. The initializer is still kept private (for now), since it relies on
some dependencies being passed down from `HUBViewControllerFactory`.

This change is made to be more compatible with Swift, where currently
`HUBViewControllerFactory` returns `UIViewController` as a type (for more
info - see https://github.com/spotify/HubFramework/issues/158).