Make HUBViewControllerScrollHandlerMock more predictable
The scroll handler mock had 2 ways of mocking content insets, one using
a property, and one using a block. This lead to a lot of confusion when
implementing tests, and could also lead to invalid test code. The property
based mocking has now been removed in favor of the block-based one.