HUBManager: Add convenience initializer
This change adds another convenience initializer to `HUBManager`, that is
the currently simplest way possible of creating an instance of the class.

Instead of having to supply their own `HUBComponentLayoutManager` and
`HUBComponentFallbackHandler`, API users can now simply supply a component
margin to use, and a block that resolves a fallback component.

To make this happen, default implementations of the above mentioned APIs
are implemented, as well as a category on `UIViewController` to make it
possible to share code between various tests.