View controller factory: Add simplest possible API for creating a VC
This change introduces a new convenience API for `HUBViewControllerFactory`,
that enables the API user to supply only the most essential information -
which content operations to use, and the title of the view controller’s feature.

The feature identifier + view URI (two other required pieces of data), are derived
from the feature title.

The reasoning behind this API is to lower the barrier for entry when starting to
use the framework, and to avoid having to supply non-interesting information in
an app that is not URI based or doesn’t use feature identifiers in any way.