Introduce HUBBlockContentOperation + context
This change introduces a content operation convenience API, that enables simpler
operations to be constructed using a block - `HUBBlockContentOperation`.

To avoid having to pass many parameters to the block, a contextual object API
is introduced (similar to how the `HUBAction` API has a `HUBActionContext`).
This contextual object API is currently only used for the newly introduced
block operations, but the plan is to migrate all `HUBContentOperations` to using
it, instead of having multiple parameters as input.