Add dynamic framework target + use it in the demo app
This will enable us to support Carthage is the proper way. The demo app
has also been updated to use the dynamic framework instead of the static
one, and to run tests against the dynamic framework. This will give us
a fairly good safety net that the static library and dynamic framework
will maintain feature pairity.

To enable tests to be run against the dynamic framework, minor changes
needed to be made to the code:

- Use `XCTAssertEqualObjects` instead of `XCTAssertEqual` to compare
  string pointers.
- Enable a bundle to be injected when loading local images, since a
  dynamic framework and its test suite do not share the same bundle.