Split running unit and UI tests
The original reason for adding the tests to the demo app’s schema was
to run the tests with the dynamic framework. This commit makes sure we
still do that on CI. However we can now control which to run. As well
as better see in the logs which steps were ran and why.

Also refactored the helper function `shouldSkipStep(step)` to be in
positive form (`shouldPerformStep(step)`). This makes it much easier to
read and reason about when used.