Rename addJSONData:error: and addJSONDictionary:
The methods where renamed to be more coherent and improve how they’re
imported in Swift. This changes the addJSONData:error: method to now
return its status and return the error via the passed in `NSError **`.

Additionally we tweak the name of the methods slightly when imported by
Swift. As to make them read nicer.

As part of this change the `HUBAddJSONDataToBuilder()` function was
refactored. Furthermore a function for setting the output error,
`BOOL HUBSetOutError(NSError **, NSError *)`, was added.

Lastly a few bugs where fixed where the output `error` was checked.
Instead of the method’s return value. This is unsafe since Cocoa
sometimes uses the passed in pointer for internal stuff.