Change all errors to use domain and code
The Cocoa standard way of creating errors is to create an error domain
that groups errors that belong together. The domain should not be used
to identify _specific_ errors.

Instead the combination of error domain and code should be used to
identify a specific error. Which should then obviously be handled
accordingly ;)

I’ve translated the existing potential errors from the unique error
domains they were using to domain + codes tuples.