View controller: Make rendering code operation oriented
This is the first in a series of tasks to make more code in the Hub Framework operation oriented.

We’ve run into a lot of issues during the last couple of months due to state management, where
things have become out of sync or required even more state to be introduced. For example, when
interfacing with `UICollectionView`, which is very asynchronous, we could end up with inconsistent
states because we would update our models too early or too late.

Operations make this a lot simpler, since using a central queue that all events pass through, you
can make more assumptions (and IMO, reason about it easier) about in what sequence various events
and bodies of work will occur.

Now, instead of having to keep state like `pendingViewModel` and `isRendering`, `HUBViewController`
now has an operation queue that it adds operations onto whenever it needs to re-render.

To make this happen, two new classes are introduced: `HUBOperation` and `HUBOperationQueue`.
`NSOperations` where considered, but ruled out due to their complexity and optimisation
towards background thread work (see the documentation for `HUBOperation` for a more in-depth
discussion). So why are these not named `HUBRenderingOperation` (+queue)? Because we want
to migrate more code to using this paradigm, but exclusively rendering code. For example,
once we move to context-based content operations (see `HUBContentOperationContext`), we could
implement a simple input/output for `HUBOperation` and use the same code for those.

But, for now, only the rendering code is updated.