Fix Hub Framework causing unsatisfiable constraints
The way Hub Framework wrapped the view of components caused UIKit
to run into unsatisfiable constraints. This affects all constraint
based views that had internal margins. For example 8 points from the
top and bottom. Since the view wasn’t given a size, and the
autoresizing mask said it couldn’t be resized automatically, such
constraints would require the view to “implode” (top edge 8 points
_below_ the view and the bottom edge 8 points _above_ the view). I.e.
an impossible layout.

This change makes sure we configure the component view such that it
will have a size. This is accomplished by setting the autoresizing
mask to be flexibile in the height and width directions.