Add first draft of an “Architectural overview”
This document is aimed at developers who are looking to contribute to the
Hub Framework, and want to get a more in-depth understanding of how the
internals work, and what patterns that were chosen (and why).

The document is currently incomplete, but still offers some value (so pushing
it up now for handover purposes). More work needs to be done to cover all
aspects of the framework (with “Hubs sessions” and the other documentation being
complementary to this document).