Fix UICollectionViewCell not preparing for reuse
The implementation of prepareForReuse is documented in
as part of UICollectionReusableView and it states that the method doesn’t
do anything. So it should be safe to not message super, right? However it
does recommend that subclasses messages super anyhow.

> The default implementation of this method does nothing. However, when
> overriding this method, it is recommended that you call super anyway.

Further the documentation notes that subclasses, such as,
UICollectionViewCell overrides the method to perform relevant actions.
Since HUBComponentCollectionViewCell subclasses UICollectionViewCell
it must message super.

> Subclasses such as UICollectionViewCell override this method and use
> it to perform relevant actions. So if your subclass descends from
> UICollectionViewCell or some other intermediate class, calling super
> ensures that your class gets the parent’s behavior.