Fix 32 vs. 64 bit issues with CGFloat
It’s a bit of a pain to work with. These functions should make it
nicer. Instead of type casting the functions makes sure to use the
correct function as provided by the OS. Depending on whether CGFloat is
defined as a double or float.

Adds the following functions:
- `CGFloat HUBCGFloatFloor(CGFloat)` – returns the value, rounded down.
- `CGFloat HUBCGFloatAbs(CGFloat)` – returns the absolute value.
- `CGFloat HUBCGFloatMax(CGFloat, CGFloat)` – return the maximum value.
- `CGFloat HUBCGFloatMin(CGFloat, CGFloat)` – return the minimum value.
- `BOOL HUBCGFloatFloor(CGFloat, CGFloat, CGFloat)` – tests whether
  two values are approximately equal (±epsilon). Caveat: it doesn’t
  work if one of the values is `0.0`.
- `BOOL HUBCGFloatIsZero(CGFloat)` – tests whether the given value is
  approximately zero.(±epsilon).
   - Should be used instead of `someFloatingPoint == 0.0` since that
     will likely yield a different result than expected. It also
     triggers a new build warning in Xcode 8.3.

Also fixes a newly introduced warning about comparing floating point
values against 0. Given that floating point values and equality checks
might yield in unexpected results. Such as:

    10.0 == 10.0 + DBL_EPSILON [yields true]

In the commit all current usage of the “normal” math variants of these
functions have been updated. There are still some places where casting
to CGFloat is necessary for the code to build for both 32 and 64 bit
architectures.

Furthermore it fixes the type (and name) of the test helper macro to
specify that it works on CGFloats.