Make HUBLiveService independent from HUBViewController
Instead of the HUBLiveService creating a view controller, have it emit
a content operation and let the implementor setup and push the view
controller. It makes the live service independent of the view
controller and leaves the user in charge of what should be done (for
example, add more content operations to the chain.